requirements.txt
# Core dependencies
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

facenet-pytorch==2.5.3

numpy>=1.24,<2.0
Pillow>=10.0.0,<11.0

# Optional but handy
tqdm>=4.66.0
